Fabric: ComponentDescriptorProviderRequest, a way to register a component on demand
Summary: This diff implements the core functionality of "reactive component registration". With a new API now we have a public API that allows registering a component by request on-demand. That can be helpful in some environments where loading of all components is undesirable or impossible due to generic app design constrains.

Reviewed By: mdvacca

Differential Revision: D17211914

fbshipit-source-id: 5a957a4f6541420d70952856599b827ceec63e3b